api-impl: ensure that SDK_INT gets passed to libandroidfw
A mismatch between SDK_INT and the sdk version that gets passed
to libandroidfw can result in confusing crashes. An `--sdk-int`
option is added to the main executable, as a shorthand
for `-X '-DBuild.VERSION.SDK_INT=<version>'`. This should make
it more straightforward to set a higher version per-app
if necessary. Additionally, `Build.VERSION.RESOURCES_SDK_INT`
is now exposed as a property, and can be used to pass
a different version to libandroidfw, although this is
discouraged.

The SDK_INT reported by default continues to be GINGERBREAD (9).